olga_telebot/requirements.txt
aiofiles==23.1.0
aiogram==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
djangorestframework==3.14.0
docker==6.1.3
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
packaging==23.2
pip==23.2.1
pydantic-core==2.6.3
pydantic==2.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
requests==2.31.0
setuptools==68.2.2
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
websocket-client==1.6.4
wheel==0.41.2
whitenoise==6.6.0
yarl==1.9.2
